TestResultCoordinator: Ignore recent results if in longhaul mode (#4455)
For longhaul, we cannot take the same approach as connectivity tests where we stop the test, allow everything in-flight to settle, then generate the result. Instead we need to be able to handle things that are in-flight, meaning some results will be unmatched.

This change allows the TRC to ignore any recently-received, unmatched expected results from the report generation. It accomplishes this through a filtering abstraction which ignores any result that arrives in the last ignore-time, with the exception of actual results matching an expected result not in the ignore time. The default ignore-time is set to 1 minute.

There are really three conceptual changes in this PR:
1. The addition of the filtering abstraction, which includes changes in the settings, report generator factory, etc.
2. The conversion of ITestResultCollection to IAsyncEnumerable. There was a TODO in the ITestResultCollection class indicating we should convert this when we were on dotnet 3.1. But the main reason I did it was because of some convenience methods this provided for my filtering abstraction. If I didn't do it I think I'd be forced to do something hacky to extend the manual implementation of IAsyncEnumerable. This seemed messy so I decided to switch to IAsyncEnumerable, which bloated the diff size.
3. The conversion of the longhaul direct method reports to not allow for sender results without receiver results. This scenario won't exist in longhaul so we can eliminate some tests/logic.

The cleanest way to make the filtering abstraction was to inject some comparer, however there was not a comparer implemented for direct methods. The reason for this is because the report generator uses a heuristic about how the receiver will only report successes. I made a direct method comparer anyway that can be used in the filter abstraction which compares results more thoroughly (includes sequence number). We should at some point add this comparer to the direct method report generators.

Co-authored-by: dylanbronson <dybronso@microsoft.com>
Co-authored-by: dylanbronson <55515325+dylanbronson@users.noreply.github.com>