Introduce the dedicated Packages Release Pipeline (#4500)
Introduce the dedicated Packages Release Pipeline

Currently our existing Azure-IoT-Edge-Core Edgelet Release pipeline is a GUI-based devops pipeline which has been working nicely for `iotedge` and `libiothsm` artifacts. Going forward starting with release/1.2-rc4, the `iotedge` and `libiothsm` are retired and replaced by the new Identity Service (IS). 

This pipeline is created to provide a support for the `aziot-edge` release artifact as a part of the new IS. We decided not the update the existing pipeline going forward because of its GUI-based property which is used across multiple branches (release1.1 & release1.2) in the release process. By updating the pipeline for release1.2 will inevitably break the release1.1 release.

Note: The RELEASE pipeline is almost identical to the BUILD pipeline in term of pipeline code base. We decide to be in favor of creating a new pipeline because
1. The version string different (we don't want the buildId in the final package)
2. We get to keep the Retention history in the pipeline itself.