Temporarily raise quickstart EdgeCA cert's expiry to 30 days.
Currently certd does not cap an issued cert's not-after to the issuer's.
This means a situation can happen where Edge Hub requests a module server cert
for 30 days and is granted one, but the Edge CA cert is valid for less than
30 more days.

In fact, this is what happens with self-signed Edge CA as configured by
`iotedge config apply` in quickstart mode.

Until certd is fixed, for RC4 we decided to do the simpler workaround of
just raising the quickstart Edge CA's expiry to 30 days.