Temporarily raise quickstart EdgeCA cert's expiry to 30 days. (#4521)
Temporarily raise quickstart EdgeCA cert's expiry to 30 days.

Currently certd does not cap an issued cert's not-after to the issuer's.
This means a situation can happen where Edge Hub requests a module server cert
for 30 days and is granted one, but the Edge CA cert is valid for less than
30 more days. EH won't renew its cert until it's 5 minutes to expiry,
but for most of that time the full cert chain will be invalid because of
the expired issuer and modules would refuse to connect to it.

In fact, this is what happens with self-signed Edge CA as configured by
`iotedge config apply` in quickstart mode before this change, because
the expiry defaults to 1 day.

Until certd is fixed, for RC4 we decided to do the simpler workaround of
just raising the quickstart Edge CA's expiry to 30 days.