Merge 1.1 to 1.1-k8s-preview (#4529)
* Fix GetModules call error using RFC3339 DateTime format. (#4293)

Fix issue calling GetModuleLogs using DirectMethod with a payload that contains parameters using RF3339 DateTime format.

* Prepare for release 1.1.0 (#4342)

* Update nuget config per security recommendations (#4289)

* Set IOTEDGE_HOST before 'make install' (#4396)

After commit a34e0ed, the iotedge CLI fails in Mariner because IOTEDGE_HOST was not set during compilation, so the CLI has the wrong path to iotedged's management endpoint.

To fix this, we set IOTEDGE_HOST prior to running `make install`. It was already being set prior to `make release`.

* Update Base Images for Security Vulnerability 3.1.12 (#4380)

- Update ARM32/ARM64 Bionic base images to 3.1.12
- Update AMD64 (Linux) Alpine base images to 3.1.12
- Update AMD64 (Windows) Nanoserver base images to 3.1.12

* Update release 1.1.0 (#4399)

Update the CHANGELOG to include the latest commits.

* Config yaml 1.1 (#4434)

This change updates the default agent tag in config.yaml to 1.1.

Cherry-pick a4faab5f62f96fa813d19a73bba0ddf92c72edf8 and update 1.1-specific Windows config.yaml.

* Upgrade sysinfo package (#4444)

This change upgrades the sysinfo package in the iotedge CLI and iotedged. It fixes bugs called out in #2747 and #3746. This change was fixed in master earlier and was supposed to by ported to the 1.1 branch, but never was.

* EFLOW: Introduce environment variables for nuget operations (#4499)

Set env variables to direct nuget caches to the data partition instead of to the rootfs

Co-authored-by: Pedro Marcelo Zara <pmzara@hotmail.com>
Co-authored-by: Damon Barry <damonbarry@users.noreply.github.com>
Co-authored-by: yophilav <54859653+yophilav@users.noreply.github.com>
Co-authored-by: ms-mahuber <60939654+ms-mahuber@users.noreply.github.com>