Fix: don't unsubscribe when there is no subscription registered (#4594)
There was an optimization on ConnectionManager so it did not do any action for subscription/unsubscription if the state did not change, to avoid excess communication to iothub. The status of a given subscription (e.g. for C2D messages) can be represented by three states: true, false or missing. Missing means that the subscription was never set, so the dictionary that stores the state has no information about it. When this happens that means that nothing has set the subscription yet, which means that no subscription exists to that topic. In this case when that subscription gets removed, that does not mean status change.

The original implementation handled this as a status change and executed a remove operation. This should not have caused any problems, however a bug in the C# SDK breaks something and it blocks subsequent subscriptions.

This fix makes the optimization work right, and it mitigates the occurrence of bug in the SDK till they fix it