Create mod.rs
What I am guessing is happening and will hopefully be able to replicate and see is, when read is at end of buffer and write is at beginning and we crash we won't get read at the right spot, instead it will be 0. We will then continue to read what is newly written and eventually write over previous data at old read pointer location. This means data loss.