Throw non-fatal exception if Mounts missing critical data.
Most data coming from createOptions is applied to the
PodSpec if set, but Mounts are a slightly different story,
it has three critical fields, and the Type and Source must
be set, or else we don't know how to handle the mount.

Creating an invalid mount in the Mounts section of
createOptions was throwing a Null References exception,
which appears to be marked as a fatal exception, and
it causes edgeAgent to crash.  A better experience is
to check the fields we need and throw a non-fatal
exception.  This also lets us report back to the user what
went wrong more clearly, doesn't crash edgeAgent, and
allows for a graceful recovery once the problem is fixed.

I did do a check othe the other mappers, and I am
reasonably certain the mappers changed in this PR are
the ones which will most likely cause NullReference
exception.