[k8s] Fix problem with multiple Binds to the same source. (#4341)
If two Binds or Mounts resolve to the same name, which implies they are the same source, reduce that to one Volume in the PodSpec.

This change keeps the behavior that the k8s Volume extensions are placed as-is, so if the user mixes volume specifications and has a naming collision, both Volumes will be put in the PodSpec, and the deployment will be reported as failed.