Fix self-signed edge-ca cert to use its subject name as the issuer name. (#4710)
34e7a6c7253c7091a9949cdccbfee93fad5761fa exposed a bug in certd that set
the issuer name of self-signed certs to the subject name from the CSR
rather than the overridden common name in the certd config. So the Edge CA
would be issued with subject name "iotedge workload ca &lt;hostname&gt;" and
issuer "iotedge workload ca", which made it a non-self-signed cert.

This was fixed in certd in
https://github.com/Azure/iot-identity-service/commit/9c7c8deeb5efefb02f1d4db37766743bf75a9d65