Import master encryption key in `iotedge config import`
The workload API exposes an encrypt and decrypt API for modules that derives
module-specific keys using the module ID and a master encryption key.

This change adds an `imported_master_encryption_key` property to
the super-config that gets set to the path of
the IoT Edge 1.1 master encryption key by `iotedge config import`.
`iotedge config apply` then copies the contents of this file to a new file
under `/var/secrets/aziot/keyd` and preloads it as the master encryption key
in the keyd config with the same key ID that edged uses.