Import master encryption key in `iotedge config import` (#4697)
The workload API exposes an encrypt and decrypt API for modules that derives
module-specific keys using the module ID and a master encryption key.

This change adds an `imported_master_encryption_key` property to
the super-config that gets set to the path of
the IoT Edge 1.1 master encryption key by `iotedge config import`.
`iotedge config apply` then copies the contents of this file to a new file
under `/var/secrets/aziot/keyd` and preloads it as the master encryption key
in the keyd config with the same key ID that edged uses.

This change also changes edged to use the master encryption key directly
to encrypt/decrypt secrets via the workload API. This was also the behavior of
IoT Edge v1.1 and earlier, but for v1.2 we changed the code to use
the master encryption key as a derivation key and
derive per-module encryption keys for this. With this commit edged uses
the master encryption key directly again.