Shorten all k8s watch timeouts. (#4618)
Watches on certain clusters seemed to be stopping and not recovering as we expected.  Investigating the code, I realized I had missed a few cases where a failure could occur, and so I fixed the code to ensure all failures were handled.  

YET, the problem still occurred.  We learned that the C# K8s client seems to have trouble with longer timeouts for List/Watch on certain clusters.  So, we shortened the timeout to below the value that seems to be where the problem lies.
Now, instead of using a default timeout, we set all watch timeouts to 3 minutes.