Cache device provisioning state (#4722)
This change caches device provisioning state on each restart. This is used to compare the device provisioning result on each restart with the previous provisioning state, so that modules can be deleted if state has changed.
This fixes the issue of iotedge removing all modules on startup, even when provisioning state was the same.