Registry API On-behalf-of calls auth check fix (#4739)
Issue: The Registry APIs exposed by the EdgeHub are used to CRUD module identities. 
For nested scenarios, special on-behalf-of APIs are supported, which allows the parent EdgeHub to CRUD module identities on behalf of the child. 
Imagine a topology like l5->l4->l3. If a module is added to l3, then the EdgeHub on l4 and l5 need to make the call to create the module on l3. For this they use the "on-behalf-of" api. Same applies to list/delete as well. 
But currently the code does not check if the caller is making the call on behalf of a descendant. This allowed, for example, l4, to CRUD modules on behalf of l5!

Fix:
Get the auth chain from the current EdgeHub to the target, and make sure the caller is on that auth chain. For example, then l4 makes a call to l5, to create a module on l3, then the auth chain from l3 to l5 is "l3;l4;l5". So l5 can validate that l4 is in the auth chain. 
Now if l4 makes a call to create a module on l5, then the auth chain from l5 to l5 is  "l5". l4 is not in this auth chain, so validation fails.