Fix ws auth with cert over ApiProxy  (#4756) (#4767)
For websockets connection the authentication needs to fail at subprotocol level to send the unauthorized to the client. 
It does the check in websockets middleware and if that fails it sets the Authenticator to NullAuthenticator. I think it is a little  weird, but otherwise would need to change the authenticator to receive the proxy token and do the proxy authentication there, better not to change it  now.