Add edgeHub identity to the scopes cache at the startup. (#4760) (#4778)
When we connect upstream, we verify that identity is in scope. This is done in order to avoid poisoning telemetry queue with the telemetry from disabled devices.

On a fresh start, we may not yet received the scopes from the upstream, so we need to force add edgeHub in the cache so it is able to connect upstream. Later in the startup process, once we get the scopes from the upstream, this record is replaced.