[mqtt] filter out publication duplicates in MessageLoader (#4745)
There is a bug in the MessageLoader which does not filter out publications after they've loaded from storage. It caused message duplicates to be sent to the $upstream sometimes out of order.

Adds a filtering step to prevent duplicates to be sent when a new batch is loaded.

Changes `fn remove(key)` -> `fn pop()` to reflect the fact we cannot remove publication by any key.