EdgeHub: Add caching to TokenProvider (#4817)
EdgeHub makes a call to the Workload API SignAsync to get a token every time it needs one. If too many such calls are made, some of them could fail. 
This code adds logic to cache a previously obtained token, and reuse it. The cache duration is half of the token expiry duration, after which it should get a fresh token.