Longhaul: Filter duplicate results in relayer (#4682) (#4831)
Currently there is a test bug where edgehub will duplicate a message from loadGen to relayer (acceptable behavior). This duplication causes the TRC to have duplicate expected results, which it is not configured to handle. The TRC will intentionally fail the tests for duplicate expected results, so this test bug results in a test failure for acceptable behavior.

The best fix I think is to filter the duplicates in the relayer test module up to some threshold. I have introduced ```DuplicateMessageAuditor``` in order to filter these duplicates.

Copied from DuplicateMessageAuditor abstraction:
```
    // Sometimes EdgeHub itself will duplicate a message when sending to this module.
    // When this happens, we don't want to report two expected results to the TRC, as
    // the TRC intentionally will fail duplicate expected results. Therefore we should
    // filter duplicate messages from edgehub up to some tolerance period. If this
    // tolerance for duplicates is exceeded, this module will not filter the proceeding
    // duplicate messages and will report duplicate expected results to the TRC, which
    // will fail the tests. That way we will know something is wrong.
    //
    // Below is how the duplicate filtering will work regarding different incoming messages.
    //
    // Case 1: New sequence number
    //         This cannot be a duplicate, so we don't filter
    // Case 2: Repeating sequence number under duplicate threshold
    //         We expect edgehub to sometimes send duplicates, so it is ok.
    // Case 1: Repeating sequence number over duplicate threshold
    //         Something weird is likely going on, so we won't filter in order to fail the tests
```

I also have a diff stacked and ready on this PR to make the TRC explicitly fail duplicate expected results so it is obvious in the test report. Another import thing to note is that this test module probably runs in our e2e test suites, but I don't expect it to cause issues.