Fix Windows construction of identity cert PKCS#12 blob to not leak keys to disk.
schannel and native-tls require a PKCS#12 blob of the device ID certificate
to use it as a TLS client cert. Before this change, the blob was created
using openssl even on Windows. However, with such a blob, schannel creates
the private key on disk under the user's persisted keys directory.
Since iotedged runs as LocalSystem, this was
C:\ProgramData\Microsoft\Crypto\RSA\S-1-5-18

It's empirically known that schannel doesn't create the disk file if
the private key is marked as "exportable". Since "exportable" is
a Windows-specific attribute, openssl doesn't have any way to set it.
So this commit uses winapi to create the PKCS#12 blob instead.

Other changes:

- Revert two uses of single-threaded tokio runtime back to multi-threaded.
  deb2753bc2e36bf2355ddb9427262e5d93d96d72 accidentally changed these while
  changing the tests.