E2E: Add regression test for custom mqtt topics (#4857)
This PR is adding an E2E test for custom mqtt topics using the existing ```genericMqttTester``` image we run in longhaul. 

This test will deploy edgehub with the broker enabled and two genericMqttTester modules. One genericMqttTester module will be in initiate mode and the other will be in relay mode. The initiate mode module will publish on an mqtt topic ```initiate/1``` and the relaying module will subscribe to this topic. When the relaying module receives a message, it will publish it back on ```relay/1```. The initiating module will be subscribed on this ```relay/1``` topic and will report to the TRC when it receives the message.

In addition to adding the test it:
1. Refactors shared logic into util abstractions for cleaner code. Specifically adding the broker/TRC to the deployment.
2. Introduces a configuration for defining the restartPolicy for test module deployed in the test. This is done so that the genericMqttTester module will no start back up after it sends its defined amount of messages.
3. Improves the genericMqttTester module to use the module name when reporting to the trc rather than some hardcoded string "genericMqttTester". This allows the dotnet test code we right here to be more understandable.
3. Updates the documentation to include new parameters required for this test