Ensure new TLS server certificate is always created on CertificateManager::new() (#4921)
The https connection between iotedged and iotedged-proxy on a cluster uses a server certificate which expires after `auto_generated_ca_lifetime_days`. When that expiration hits, iotedged is supposed to renew that certificate, but the HSM API simply returned the existing (expired) cert.  This cascades into edgeHub eventually losing its server certificate, and the whole runtime is disconnected.

This PR changes the method for acquiring a new certificate to always delete the existing certificate by alias.