Fix getDeviceAndModuleOnBehalfOf to check if target device is in scope (#4886)
Issue: EdgeHub allows child EdgeHubs to get device/module information for the devices in their scope (l3/$edgeHub can ask for l3child information to l4/$edgeHub). But a child EdgeHub should not be able to ask for information on identities that are not its descendants. But the current code did not do this check. Consequently a child edgeHub (say l3/$edgeHub) can ask for information on an identity that is not its child (say l4child) and be able to receive it.

Fix: Before returning the information to the requesting child, also check if the target device/module is one of its descendants.

Chery-pick: https://github.com/Azure/iotedge/commit/5e1028ec8db3b79c61969abeffe1b2a3701adf8a