Update functions to 3.0 (#4910)
This reverts commit 3ab67a3a199f1bb47123f59f3310eb902f13f823.
Filters out bin folder when running tests to fix the tests failure issue. It's not clear why only in some runs the bin folder has the test dlls and not the runtimeconfig.json files which is causing the tests run to fail. Other times it was passing because the bin folder didn't contain the test dlls when running on different pipeline agent (but using same dotnet sdk and everything else looked the same)