obsagent-client test module (#4931)
This PR adds an Edge module that sends example metrics to a configurable OTLP endpoint. The specific endpoint this module is intended to connect to is that of the Edge observability-agent. However, this module also successfully connects to the OTel Collector (written in Go).

This obsagent-client module loops, updating and exporting metrics for one example instance of each OTel instrument type (i.e. counter, up-down-counter, value-recorder, sum-observer, up-down-sum-observer, and value-observer). The rate at which the metrics are updated is configurable. The rate at which the metrics are exported (or pushed) out of the module is also configurable.

This test module can be configured via the following environment variables:

1. UPDATE_RATE - Rate at which each instrument is updated with a new metric measurement (updates/sec)
2. PUSH_RATE - Rate at which measurements are pushed out of the client (pushes/sec)
3. OTLP_ENDPOINT - Endpoint to which OTLP messages will be sent.