Fix: overwrite 'IsDirectConnection' flag when device changes from Indirect (#4977)
When a device connects indirectly, then disconnects and connects directly again, the following happens:
As we don't have "indirect disconnection", the device proxy instance lingers for a while before gets deleted by idle-time. Because of that it may show up in the "knownIdentities" list and the way added connections are calculated (comparing the already known connections with the broker sent connection list), it will not be added as a "direct connection" but keeps existing as "indirect connection".
As a result, the device proxy will use mqtt topics for indirect connections which will not be subscribed by the directly connected device. As a solution, check if the broker sent connection-list and the known connections match by 'IsDirect' flag. If not, remove the indirect connection and recreate them as direct.