Restore device scopes from older store (version < 1.2) (#4889) (#4985)
cherry-pick #4889 

Edgehub version 1.1 didn't have DeviceScope and ParentScopes in storage and updating to 1.2 fails to start edgeHub because reading from cache fails. 
Fix is to check the service identity that was read from storage for current device (from actorDeviceId) and if the DeviceScope is empty that means it is a restore from older version (edge devices should always have DeviceScope) and set the value for DeviceScope from device details (deviceId and generationId). For leaf devices it sets DeviceScope and ParentScopes to the edge device scope because if the device was in store it must be a child of the edge device and sync from cloud will update the leaf device in case it was removed as child.

This fixes #4828