Fix edgehub queue len metric (#4952)
This attempts to fix the issue of queue len metric reporting incorrectly due to how MessageStore calculated the current queue length. This change now gives the ability to directly request the Count of items from the store implementation.

There is also a matter of discussion on whether we can get the metric to exist solely in one unified place and whether we require both Checkpointer and MessageStore.