New Helm Chart (0.2.13), new image tag (1.1-k8s-beta15) (#5031)
Update Helm charts and images for beta15
New in Beta 15:

[k8s] Updating native-tls will allow root ca certs in trust bundles. (#4980)

If the customer had a trust bundle containing root CAs already in the certificate store,
creating a TLS connection would fail. This will avoid that error.

Add iotedged.data.runtimeLogLevel to helm chart settings. (#4936)

The iotedged pod had a default logging value which could not easily be changed.
This allows the customer to set it through the Helm charts.

Ensure new TLS server certificate is always created on CertificateManager::new() (#4921)

The iotedged pod was not correctly renewing its server certificate. This caused the edge runtime
to fail after 90 days (the default timeout for the server cert).

"task" is awaited twice in KubernetesEnvironmentOperator. This may have the effect of silently shutting down the Environment operator when the "ListPods" fails in specific ways.(#4793)

Removed what looks like a merge error.

cherry-pick all 1.1 branch commits.