Fixes setting up env var when receiving new config (#5053)
This issue was found while troubleshooting with ACR.
New config would be parsed but the environment variables used to resolve the config are determined statically at start up.

This is an issue is the config is reloaded with different env var than the original config.

Test:
Tested with ACR config and with our own default config.