Add "dist" target to supply Yocto builds with release. (#5057)
Yocto builds are currently using a "vendor" file - a tar-gzipped snapshot of the project created by `cargo vendor` - to build the Yocto images, and it's getting it from the release GitHub page.  As a first step to improve the process to update Yocto, I'm adding this to the packages build. This only needs to be done once, so I attached it to the ubuntu18.04 build, but it could be any build.