Don't dispose stream too early. (#5097)
In my previous pr #5035 I tried to optimize disposing of UDS. That fix was incorrect, since we rely on the fact that Stream is still open when we return `HttpResponseMessage` from `SendAsync`. Later, the consumer of `SendAsync` will dispose the `HttpResponseMessage` and it will in turn dispose the `Stream` and `Socket`.