Fixed exception type in BrokerConnection::SendAsync,  (#5091)
so operation will be retried if no mqttclient exists.

The problem occurred when EdgeHub was shutting down but a message was being sent upstream. As the protocol head was stopped, that cleared the mqttclient client, so the send operation in SendAsync() threw an 'Exception' instance, which was not retriable by CloudEndpoint, so it just skipped the message.