Devops Sync Job: Refactor to facilitate incoming bug opening automation changes (#5122)
For the bug opening changes, it will be cleaner if we refactor the sync job to promote cleaner modifications.

Currently the reverse callstack for the relevant part of the sync job works like this:
```
- RunAsync
  - foreach branch
    - ImportVstsBuildsData
      - ImportVstsBuildsDataForSpecificDefinition
        - GetBuildsAsync(definition)
        - UpsertToDB
        - Get the most recent build timestamp, track this, then pass it all the way up the callstack to store in double hashmap
```

It will be cleaner if we get the builds and update the most recent timestamps upfront in some encapsulated function. Then we can pass the list of bugs into the ImportVstsBuildsDataForSpecificDefinition and whatever function I create in a future PR to automatically log bugs. Here is what the new reverse callstack looks like:
```
- RunAsync
  - foreach branch
    - foreach buildDefintionId
      - helper function for getting builds and updating most recent timestamp
      - ImportVstsBuildsDataForSpecificDefinition
        - UpsertToDB
      - LogDevopsBugForFailingBuilds *TODO
```

I also noticed that the areas for the dashboard bug chart were missing some categories so I added them.